+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # [develop](https://github.com/adhearsion/adhearsion-ivr)
   * Feature: Added support for input options to be passed to ask
+  * Feature: Added support for output options to be passed to ask
 
 # [v0.2.0](https://github.com/adhearsion/adhearsion-ivr/compare/0.1.0...0.2.0) - [2015-03-27](https://rubygems.org/gems/adhearsion-ivr/versions/0.2.0)
   * Feature: Added support for defining alternative prompt resolution algorithms by overriding `#prompts`